refactor: move label to widgets folder
