+++ selfdrive/ui/onroad/alert_renderer.py
@@ -2,12 +2,12 @@ import time
 import pyray as rl
 from dataclasses import dataclass
 from cereal import messaging, log
+from openpilot.selfdrive.ui.ui_state import ui_state
 from openpilot.system.hardware import TICI
 from openpilot.system.ui.lib.application import gui_app, FontWeight, DEFAULT_FPS
-from openpilot.system.ui.lib.label import gui_text_box
 from openpilot.system.ui.lib.text_measure import measure_text_cached
 from openpilot.system.ui.lib.widget import Widget
-from openpilot.selfdrive.ui.ui_state import ui_state
+from openpilot.system.ui.widgets.label import gui_text_box
 
 ALERT_MARGIN = 40
 ALERT_PADDING = 60
