+++ selfdrive/ui/onroad/driver_camera_dialog.py
@@ -5,7 +5,7 @@ from openpilot.selfdrive.ui.onroad.cameraview import CameraView
 from openpilot.selfdrive.ui.onroad.driver_state import DriverStateRenderer
 from openpilot.selfdrive.ui.ui_state import ui_state
 from openpilot.system.ui.lib.application import gui_app, FontWeight
-from openpilot.system.ui.lib.label import gui_label
+from openpilot.system.ui.widgets.label import gui_label
 
 
 class DriverCameraDialog(CameraView):
