+++ system/ui/updater.py
@@ -7,10 +7,10 @@ from enum import IntEnum
 
 from openpilot.system.hardware import HARDWARE
 from openpilot.system.ui.lib.application import gui_app, FontWeight
-from openpilot.system.ui.lib.label import gui_text_box, gui_label
-from openpilot.system.ui.lib.wifi_manager import WifiManagerWrapper
 from openpilot.system.ui.lib.widget import Widget
+from openpilot.system.ui.lib.wifi_manager import WifiManagerWrapper
 from openpilot.system.ui.widgets.button import gui_button, ButtonStyle
+from openpilot.system.ui.widgets.label import gui_text_box, gui_label
 from openpilot.system.ui.widgets.network import WifiManagerUI
 
 # Constants
